all dummy data deleted

+++ app/src/main/java/com/example/paulina/wisehome/model/businessobjects/WeatherMeasurements.kt
@@ -6,4 +6,4 @@ import java.io.Serializable
 class WeatherMeasurements(
         val temperature: Long,
         val humidity: Long,
-        val createdAt: Long) : Serializable
+        val createdAt: Double) : Serializable
